requirements.txt
numpy>=1.19.2
pillow
cython
matplotlib>=3.3.4
opencv-python>=4.5.1
scikit-image>=0.18.1
pandas>=1.2.4
seaborn>=0.11.1
scipy>=1.6.2
shapely
torch>=1.8.0
torchvision>=0.9.0